feat: implementação das rotas de categoria

+++ src/services/category/deleteCategory.service.ts
@@ -1,15 +1,19 @@
 import { AppDataSource } from "../../data-source";
 import { Category } from "../../entities/category.entity";
+import { AppError } from "../../errors/appError";
 
 const deleteCategoryService = async (categoryId: string) => {
   const categoryRepository = AppDataSource.getRepository(Category);
   const categories = await categoryRepository.find();
 
-  const category = categories.find((category) => category.id === categoryId);
+  const category = categories.find((elem) => elem.id === categoryId);
+
+  if (!category) {
+    throw new AppError(404, "No category with this id");
+  }
 
   await categoryRepository.delete(category!.id);
-
-  return true;
+  return category;
 };
 
 export default deleteCategoryService;
